Fix mutex implementation
* Reentrant SMP-able mutex
* Correctly initialize all mutexes
* Fix hashing for hashmap

+++ kernel/include/kernel/tasking/wait_resolver.hpp
@@ -26,7 +26,7 @@
 
 struct g_wait_resolver_sleep_data
 {
-	uint64_t wakeTime;
+	uint32_t wakeTime;
 };
 
 bool waitResolverSleep(g_task* task);
